requirements.txt
# Core framework
langgraph==0.2.39
langchain==0.3.26
langchain-core==0.3.66
langchain-community==0.3.26
langchain-openai==0.3.25
langchain-google-genai==2.1.5
langchain-anthropic==0.2.4

# Data models and validation
pydantic==2.9.2
pydantic-settings==2.6.1

# HTTP requests and RSS parsing
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3

# Database
supabase==2.8.0

# Vector operations and similarity
faiss-cpu==1.7.4 ; platform_system=="Linux"
openai==1.54.4
numpy==1.26.4
scikit-learn==1.5.2

# Text processing
python-dateutil==2.9.0.post0
pytz==2024.2

# Template engine
jinja2==3.1.4

# Logging and monitoring
structlog==24.4.0
langsmith==0.1.140

# Environment and configuration
python-dotenv>=1.0.0
click==8.1.8

# OGP image generation (for Phase 2+)
playwright==1.48.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Utility
tqdm==4.66.5

# Markdown preview generation
markdown==3.5.2

# Newly added
hdbscan